+++ lib/warpath.ex
@@ -11,6 +11,77 @@ defmodule Warpath do
   alias Warpath.Execution.Env
   alias Warpath.Expression
 
+  @type json :: String.t()
+  @type container :: map | list
+  @type document :: container | json
+  @type selector :: Expression.t() | String.t()
+  @type opts :: [result_type: :value | :path | :value_path | :path_tokens | :value_path_tokens]
+
+  @doc """
+  Remove an item(s) from a nested data structure via the given `selector`.
+
+  If the selector does not evaluate anything, it returns the data structure unchanged.
+
+  ## Examples
+
+      iex> users = %{"john" => %{"age" => 27, "country" => "Brasil"}, "meg" => %{"age" => 23, "country" => "U.K"}}
+      ...> Warpath.delete(users, "$..age")
+      {:ok, %{"john" => %{"country" => "Brasil"}, "meg" => %{"country" => "U.K"}}}
+
+      iex> numbers = %{"numbers" => [20, 3, 50, 6, 7]}
+      ...> Warpath.delete(numbers, "$.numbers[?(@ < 10)]")
+      {:ok, %{"numbers" => [20, 50]}}
+
+      iex> users = %{"john" => %{"age" => 27}, "meg" => %{"age" => 23}}
+      ...> Warpath.delete(users, "$..city")
+      {:ok, %{"john" => %{"age" => 27}, "meg" => %{"age" => 23}}} # Unchanged
+
+  """
+  @spec delete(document(), selector()) :: {:ok, container()} | {:error, any}
+  def delete(document, selector) do
+    case query(document, selector, result_type: :path_tokens) do
+      {:ok, paths} ->
+        data =
+          paths
+          |> maybe_wrap()
+          |> do_delete(document)
+
+        {:ok, data}
+
+      error ->
+        error
+    end
+  end
+
+  defp maybe_wrap(paths) do
+    case paths do
+      [{:root, _} | _] -> [paths]
+      _ -> paths
+    end
+  end
+
+  defp do_delete(paths, document) do
+    paths
+    |> Enum.uniq()
+    # Remove highest index first
+    |> Enum.sort(&>=/2)
+    |> Enum.reduce(document, fn path, data ->
+      data
+      |> pop_in(to_accessor(path))
+      |> elem(1)
+    end)
+  end
+
+  defp to_accessor([{:root, _} | path_tokens]) do
+    Enum.map(path_tokens, fn
+      {:property, property} ->
+        property
+
+      {:index_access, index} ->
+        Access.at(index)
+    end)
+  end
+
   @doc """
   Query data for the given expression.
 
@@ -36,11 +107,6 @@ defmodule Warpath do
     * `:path_tokens` - return the path tokens instead of it string representation, see `Warpath.Element.Path`.
     * `:value_path_tokens` - return both value and path tokens.
   """
-  @type json :: String.t()
-  @type document :: map | list | json
-  @type selector :: Expression.t() | String.t()
-  @type opts :: [result_type: :value | :path | :value_path | :path_tokens | :value_path_tokens]
-
   @spec query(document, selector(), opts) :: {:ok, any} | {:error, any}
   def query(document, selector, opts \\ [])
 
@@ -116,4 +182,40 @@ defmodule Warpath do
     do: {member, Enum.reverse(path)}
 
   defp collect(%Element{value: member}, _), do: member
+
+  @doc """
+    Updates a nested data structure via the given `selector`.
+
+    The `fun` will be called for each item discovered under the given `selector`, the `fun` result will be used to update the data structure.
+
+    If the selector does not evaluate anything, it returns the data structure unchanged.
+
+  ## Examples
+      iex> users = %{"john" => %{"age" => 27}, "meg" => %{"age" => 23}}
+      ...> Warpath.update(users, "$.john.age", &(&1 + 1))
+      {:ok, %{"john" => %{"age" => 28}, "meg" => %{"age" => 23}}}
+
+      iex> numbers = %{"numbers" => [20, 3, 50, 6, 7]}
+      ...> Warpath.update(numbers, "$.numbers[?(@ < 10)]", &(&1 * 2))
+      {:ok, %{"numbers" => [20, 6, 50, 12, 14]}}
+
+      iex> users = %{"john" => %{"age" => 27}, "meg" => %{"age" => 23}}
+      ...> Warpath.update(users, "$.theo.age", &(&1 + 1))
+      {:ok, %{"john" => %{"age" => 27}, "meg" => %{"age" => 23}}} # Unchanaged
+  """
+  @spec update(document(), selector(), (term() -> term())) :: {:ok, container()} | {:error, any}
+  def update(document, selector, fun) do
+    case query(document, selector, result_type: :path_tokens) do
+      {:ok, paths} ->
+        data =
+          paths
+          |> maybe_wrap()
+          |> Enum.reduce(document, &update_in(&2, to_accessor(&1), fun))
+
+        {:ok, data}
+
+      error ->
+        error
+    end
+  end
 end